parsePDFs01/requirements.txt
pip
pandas
tabula-py
openpyxl
Jpype1